Separated the Filter and Sort in Own File
